# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ContactModel": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createDateColumn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"telephone",
					"email",
					"createDateColumn",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IContactPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"telephone",
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DeleteResult": {
				"description": "Result object returned by DeleteQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Number of affected rows/documents\nNot all drivers support this"
					}
				},
				"required": [
					"raw"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/contact": {
			"get": {
				"operationId": "GetContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ContactModel"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContactModel"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IContactPayload"
							}
						}
					}
				}
			}
		},
		"/contact/{id}": {
			"get": {
				"operationId": "GetContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ContactModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ContactModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IContactPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteContact",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResult"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}